Implement a DNS resolver on my vacation
Knowing that I would not have internet on my laptop but still have the
ability to hit a DNS server from the payment gateway, I thought it would
be fun to download a few key RFCs and implement from scratch. I allowed
myself a university article "hello world" refrence for making an A
record request, but quickly switched to RFCs to debug and triage.

I was able to use dig to compare responses but I didn't allow myself the
ability to use wireshark (or termshark) because I forgot to download it
before I left.

This supports requesting A, MX, AAAA, and SOA records. It can decode the
same four types but it does not implement all resource types. It will
however print out the RDATA of a record type that it does not know
about. I wasn't able to find a DNS test harness once I left to ensure
the decoding of the other message types was accurate.